Allow different stores for the state of the circuit breaker
Usually you have several app servers running multiple processes, so
there should be a way to synchronize all of them.

I like how focused this gem is, and I agree, it shouldn't be in it's
scope to provide fixed store like redis. Instead I think it should be
possible to provide another backend (batteries not included.)